services/workspace/requirements.txt
# Framework
fastapi>=0.115
uvicorn[standard]==0.30.0
httpx==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
psycopg2-binary>=2.9
alembic>=1.13
pgvector>=0.3

# Auth
python-jose[cryptography]>=3.3
bcrypt>=4.0

# Redis
redis>=5.0

# Export
openpyxl>=3.1

# Email
resend==2.26.0

# Google Sheets MVP
google-api-python-client>=2.169
google-auth>=2.40
